Add a proof-of-concept for rootless operation
The existing script works by detaching the mouse's kernel driver and
sending raw USB commands to change the LED colors.  This has some
downsides:
- Mouse briefly stops working.
- Root privileges required every time colors are changed.
- PyUSB dependency.

This new script operates on a different principle.  It sets the proper
values on Linux's internal HID report using the hiddev driver and allows
the kernel to translate that into the equivalent USB command.  This
addresses the downsides of the previous approach:
- Mouse stays connected the entire time, since it uses a separate
  USB interface from the LEDs and we never detach kernel drivers.
- Don't need root to change colors.  Access is controlled by the hiddev
  device.  We need root access one time to install a udev rule that
  controls the permissions and installs a handy symlink.
- No external dependencies (though we do have to translate a lot of
  definitions from the Linux C headers into Python).

This is not a fully-working solution: just a proof-of-concept that
allows setting a solid color.